Add classes & functions for all plugin layers
The plugin effectively has three layers - app-facing, interface and
platforms.  Here we completed the classes and glue functions for each
layer so that everything from Box2D is available in the app-facing layer.

The architecture we have chosen for now makes the plugin class
(FlutterBox2DPlatform) a factory  that can be used to create all of the
delegates (the platforms specific implementations).  We also create
classes in the app-facing layer, that wrap each delegate.  When a new
class is created it's delegate is created behind the scenes. This
pattern allows for a cleaner API at the app-facing layer.

Also, B2PolygonShape "is a" B2Shape which in the app-facing layer is
handled by B2Shape being an abstract class with no _delegate but a
getter named _delegate.