feat(b2dp): add platform package for macos
I added the platform package and wired it up with the app-facing and
interface packages of the plugin. I also added a mac target to the
plugin's example to test the new platform package.

In order to get the Box2D library compiled into the app I added Box2D.a
(compiled with a clone of the original repo) along with the headers and
to macos/box2d in the mac platform package and edited it's podspec to
include the files in the build.