feat(fb2d): Add mac platform package (#60)
* feat(b2dp): add platform package for macos

I added the platform package and wired it up with the app-facing and
interface packages of the plugin. I also added a mac target to the
plugin's example to test the new platform package.

In order to get the Box2D library compiled into the app I added Box2D.a
(compiled with a clone of the original repo) along with the headers and
to macos/box2d in the mac platform package and edited it's podspec to
include the files in the build.

* Remove box2d C++ and static lib & add basic C++

* Add Box2D headers, bindings generated with ffigen

I added some C functions to native_add.cpp that are taken from
emscripten's bindings generator, which breaks the C++ classes out into
C functions, in order to test calling into the static library.

I used ffigen on the C functions to generate bindings so I knew what
the method signatures should look like.