Get example running for macos
The aim was to get a basic example running so I added the relevant C
bindings, native (FFI) delegate classes & the top level Dart classes to
allow for creating a body, applying some forces and stepping through
the simulation a few times.

While trying to solve an issue with Dart finding the symbols I added the
C++ source files to the project and updated the podspec to compile Box2D
during the build. It turned out not to be the cause of the problem but
using the source to compile Box2D during the build has greater
flexibility, eg. making it easier to debug.