Add an adapter layer so JSInterop can be delegates
I reset the JSImpl classes to uses the exact method names from the JS,
as JS interop renaming only works on classes, not memebers. I also added
an adapter layer so we can have a Delegate layer with an API that works for
both FFI and JS Interop classes and uses Dart default syntax without any
JS style nonsense.

I added some notes on the JS that we are interroping with and a copy of
the file that has the JS.  The file is "box2d_glue.js", it was created
by emscripten along with a C++ glue flie, so that both files could be
used in the last step of building the WASM.