feat(fb2d): add ffigen package (#63)
I added a package to the flutter_box2d folder to sit alongside the
plugin packages. Currently it is only used for experimentation and
validation by running the ffigen on small code samples. The aim is to
eventually use ffigen and other forms of code gen to generate most of
the plugin's code.