feat(fb2d): Add rendering and a game loop
I added a GameWidget that holds a GameLoop and CustomPainter - the
CustomPainter repaints on each tick so we can render the game world.

A GameWorldPainter takes a GameWorld object and paints everything in the
world, which means our GameWidget has everything needed render the game.

I used the Update Method pattern, so the GameWorld and SquareObject have
update(dt) and a paintOn(canvas) functions.