Add a Context enum & fix arg types issue
The way 'dart_arg_types' was being calculated (via 'map' iterating over
collections and calling a function on each item) wasn't working because
the function being called expected to take the full list of interfaces
in order to check the sig against all interfaces. So I used a lambda
instead and got the desired outcome.

I also added an enum, Context, that is used to make 'type_to_dart'
behave differently in different contexts, eg. add 'Platform' to the
types when we are creating the delegates.