feat(fb2d): add inheritance for super/subclasses (#80)
I edited the Box2D.idl file to fix errors caused by b2Joint subclasses
missing the Dump() function.  For now I just added the function to each
subclass in the Box2D.idl file, later on we should figure out why the
code gen isn't generating the subclasses with the Dump() function.

I added inheritance for each of the super/subclass pairs described in
the Box2D.idl file. The inheritance relationships need to be added for
all layers: decorators, delegates and adapters, as each layer uses the
layer below as members.

I added a diagram of the class inheritance, taking the diagrams from the
Box2D docs and pasting into box2d.dio. The purpose was to have an easily
accessible reference of the inheritance relationships in the Box2D
classes.