Move to staticInterop API for JS Interop
I changed the JSImpl classes to use the new staticInterop API to try and
fix the inheritance issue we were having and it turned out to be a lot
better in several ways - inheritance works, both constructors and
functions can be renamed.

I used nullable type for function/constructor args in
overloads - only for the later args, which are optional.  This is part
of the steps to make functions with optional args in the decorator
classes.

I cleaned up the example a bit to make the code more readable.

I edited the expose-box2d.js file to expose b2CircleShape along with the
already exposed types, as the example uses the class.

I moved to the Flutter beta channel to update the js package to the
latest version, which has support for the new staticInterop API. I also
generated a new Flutter project and merged the index.html into this
project as we were getting an error on startup that seemed to come from
the framework. The issue went away.