Remove ffi_plugin_template from experiments
Reason: The experiment was successful and gave us a way forward so is no
longer needed. The problem we were facing was that the Box2D code was
not being built and no Pod was added when the plugin was used in an app
(in this case the example app). After making a new plugin with the
plugin_ffi template (while on the beta channel) and dropping in all the
Box2D C&C++, running the example app shows the Box2D code being
compiled. The next step is to step by step replace our plugin with the
latest template and undo the changes to keep our code that was added
to the initial generated plugin.