Update code gen script to move generated files
We have a script that generates bindings into code_gen/out/ then
copies the relevant files to the package where they live. The generated
just cause lots of errors where they are because they are out of place
so the  imports are wrong. I initially added ignore statements to the
analysis_options but there doesn't seem to be any reason to keep the
extra copy of the generated files.