feat(gla): use nonce to find doc & add gatherId
I updated the gather route of gather_link_account_shelf to query for the
document with the given nonce, then update that document with the
gatherId that was passed in the query parameter of the route.

I added a Document interface and DocumentGoogleApis class that
implements the interface with a named constructor that takes a
googleapis Document.

I changed the firestore service to return Documents where approriate,
rather than JsonMap.

I updated Locate to keep a client member so the client used to create a
Firestore service can be closed. We need to be able to close the client
when it is no longer needed or weird bugs will cause weird pain.