Fix & update firestore_api_client
I moved our notes on how to run the example to the parent project's
README so everything is in one place.

I also added a use case to the example - searching for documents
with a certain key. The example still needs to be fleshed out and have
a range of use cases added.

I added a fix to getDocuments to remove all items from the response with
null document member so we can run through all items in the response and
assume no nulls.