Setup gather_link_account_shelf to run locally
I added firebase config files via `firebase init` so we can run the
emulator.

I added a `provideLocalFirestore` function to the `Locate` class to provide a
local firestore, which creates a `GoogleapisFirestoreService` that is
pointed at the local firestore emulator.

I improved the error handling by adding a stack trace to the data added
to the firestore on error.