feat(gla): use nonce to find doc & add gatherId (#118)
* feat(fsi): add a Document interface

I added an abstract class that firestore_service_googleapis and
firestore_service_flutterfire will implement.

* feat(gla): use nonce to find doc & add gatherId

I updated the gather route of gather_link_account_shelf to query for the
document with the given nonce, then update that document with the
gatherId that was passed in the query parameter of the route.

I added a Document interface and DocumentGoogleApis class that
implements the interface with a named constructor that takes a
googleapis Document.

I changed the firestore service to return Documents where approriate,
rather than JsonMap.

I updated Locate to keep a client member so the client used to create a
Firestore service can be closed. We need to be able to close the client
when it is no longer needed or weird bugs will cause weird pain.

* Fix & update firestore_api_client

I moved our notes on how to run the example to the parent project's
README so everything is in one place.

I also added a use case to the example - searching for documents
with a certain key. The example still needs to be fleshed out and have
a range of use cases added.

I added a fix to getDocuments to remove all items from the response with
null document member so we can run through all items in the response and
assume no nulls.

* Setup gather_link_account_shelf to run locally

I added firebase config files via `firebase init` so we can run the
emulator.

I added a `provideLocalFirestore` function to the `Locate` class to provide a
local firestore, which creates a `GoogleapisFirestoreService` that is
pointed at the local firestore emulator.

I improved the error handling by adding a stack trace to the data added
to the firestore on error.