Make button work, make button nice
I had to add the... accountID?... to the index.html. rtfm

There is a bug in the latest release of firebase_core_web so I
downgraded. The fix is already in flutterfire repo so we can undo the
fuggery after next release.

I added a google image and played around with the button style... mwa!