feat(lcr): add locator packge
I made a simple service locator for use with a new redux package
I'm working on.

The goal for the locator package is to add features to support more
use cases over time but this is just a minimal starting point, to get
something up to github so external packages can depend on locator.