feat(tpr): Complete AppBar feature
I made a custom AppBar called BasicAppBar because it's simple to implement
and we don't need most of what AppBar does.

I re-created the macos target because it was a bit old and behaving
poorly. The new fresh target seems to be a happy camper. I left out the
prebuilt Firestore thing in the podfile b/c it's less important now that
the default Flutter setup does some pods build artifact caching and
the prebuilt setup needs to be kept updated and sometimes causes
problems.

I added smoke tests that the app bar gets built by it's parent and
builds it's children.