new(fai): add firebase_auth_service_interface
I added a package that provides an interface for different
implementations of a FirebaseAuth service.  redfire already has such a
service, that uses flutterfire's firebase_auth package.

The reason for the interface is so that we can add other
implementations of a firebase auth service and use them interchangeably
where possible or just a single API surface to remember between
platforms.

Other implmentations that may be useful are one that uses
googleapis_auth or one that use the firebase auth REST API.