Remove vm_service dependency & clean up
I added a model: ReduxStateEvent, for an event in a stream that carries
state change information, and I renamed EventsModel to StateChanges,
which better represents what the model is.

I renamed RedFireStatefulWidget to ReduxDevToolsScreen which is a more
accurate name for what the widget is & does.

I imported json_utils, just for the JsonMap typedef.