feat(rfr): Add Redux DevTools Screen
The redux_devtools_screen package was extracted from the Redux DevTools
Plugin, which was being developed in a fork of the Flutter DevTools
(as the plugin architecture has not arrived just yet).  This was done so
we can have a common package shared by the plugin and redfire.

In redfire we put the ReduxDevToolsScreen alongside the MainScreen with
a SplitView so the DevTools is visible while we interact with the app.

I added a global stream controller which is used in the
DevToolsMiddleware (to send events) and in the ReduxDevToolsScreen
(to receive events). In a future PR I'll add our new locate package and
and change the global to be in a service locator.