feat(rfr): Only build redux-devtool-screen on flag
I extracted the StoreConnector that was updating the Navigator whenever
AppState's pages list changes, into a widget called PagesNavigator.

The AppWidget now only builds a SplitView with ReduxDevToolsScreen
and the PagesNavigator when a constant called "REDUX_DEVTOOLS"
is found in the environment. If no such constant is found, just the
PagesNavigator is built.