Add Sync and Async actions, refactor & clean up
I broke apart the single file in redaux, to follow best practice
(or perhaps it's just my preferred style).

I removed the ReduxService type from redaux as there doesn't seem to
be an advantage of services having the same supertype but there is a
disadvantage of coupling them to redaux.

I removed the 'Action' from the actions and made some notes on style
choice in the docs. It kind of just feels cleaner.

The user defined actions now inherit from either AsyncAction (which
 has a Middleware member) or SyncAction (which has a Reducer member)

 I added a 'history' member to the Action super-type to hold
 AsyncActions, which will be used to keep a record of the sequence of
 actions resulting in a state change, for better visualisations in
devtools.