refactor(+): Add RootState, safeAsyncCall
I changed State to `RootState extends State` so that we can distinguish
between the state items in the redaux state tree and the root node of
that tree. Eg. so we can add a errorMessages item to the rootnode,
meaning all `AppState extends RootState` will have errorMessages.

I added a `safeAsyncCall` method to the Store which wraps each
middleware call in try/catch inside an async function and adds
errorMessages to the AppState as needed.

`middleware` and `reducer` members (of Async and Sync actions) are no
longer nullable, as the nullable types was a left over from a previous
architecture and isn't needed any more.