new(jst): Add the json_types package (#151)
This is just the refactored json_util package - the next PR has
details on why and how, this PR is to put up a package so
version solving finds a package when it looks in the git path,
which seems to happen when a dependency in turn uses a dependency
with a git path, even when there is an override. The path dep does
seem to eventually be used but we need a package in the github repo
to keep the solver happy.