rdx: Add toJson and emit DispatchEvents
I moved ErrorMessage into a separate file in models directory, for
better project organisation and to follow our style guide.

I added a toJson to Action interface so that the Store can call it if in
DEVTOOLS mode and sending each DispatchEvent to the devtools
screen. I also added a parentId member to hold the 'parent' action so
devtools can use that information to be helpful when providing
visualisations.

The State interface now has a toJson method for the same reason as
Action, so that the Store can call it when sending DispatchEvents to the
DevToolsScreen.

I renamed _streamController to _stateChangesController and added
_dispatchEventsController, which emits json for the action and
(potential) state change on each action dispatch.  The state change
events are for widgets and the dispatch events are for DevTools.