Add data structures to map actions to lineage info (#158)
I added a `parentId` parameter to the abstract Action.toJson so that
a parent id can be sent with the json for an action but the parentId
(which is only used in devtools) isn't a member of Action.

In the Store I updated the action.toJson call to add the parentId,
if present.

I updated all the actions in our_meals as well as the snippets with
the change of adding an optional parentId parameter to the toJson
function.

I added an `_indexFor` map and as each new DispatchEvent is added,
the actionId is mapped to the list index that it has been given.

I added an `identifyLineage` to the DispatchEvents model that is called
when an item is selected.  The function follows the parent chain, using
the `_indexFor` data structure that maps actionId to list index, and
adds lineage information to another map, which can in turn be used by
each ActionHistoryItem to get it's lineage info.