Add lineage arrow by adding shapes to each item
I updated the UI for the ActionsHistoryItem to match the designs and
added drawing of the LineageShape, which is an the arrow that shows
the lineage of the selected action, so we can easily see how the
selected action came to be.

I moved the logic of selecting the widgets to show for lineage out of
the build and instead have each enum hold the relevant widgets. The
result is a lot cleaner, more readable code.