chore(+): Update RDS state management to redaux  (#159)
* Small fixes to DispatchEvents algorithm

I updated identifyLineage() to calculate and set correct lineage shapes
for each action, following each parent pointer from the selected action
to the originating action, then iterating over the list between origin
and endpoint to set the remaining actions to the 'notConnection' shape
(which runs the lineage arrow behind the list item.

* Add design docs & ubiquitous language

* Add lineage arrow by adding shapes to each item

I updated the UI for the ActionsHistoryItem to match the designs and
added drawing of the LineageShape, which is an the arrow that shows
the lineage of the selected action, so we can easily see how the
selected action came to be.

I moved the logic of selecting the widgets to show for lineage out of
the build and instead have each enum hold the relevant widgets. The
result is a lot cleaner, more readable code.

* Clean up - upgrade deps, remove redundant code

* Add comments to DispatchEvents model

* Rename redux_devtools_screen

Renamed to redaux_dev_tools_screen to match our style guide (because
the class name was ReduxDevToolsScreen) and changed to redaux b/c the
DevTools are now visualising things that are specific to Redaux (like
the action lineage).

* Convert to redaux for state management

I also did some refactoring of folders and file/class names for improved
clarity & readability.

* Fix DevToolsState to not call non-existent super

* Improve message in StoreProviderNotFoundError

* Clean up commented out code

* [RDX] Update stateChanges stream to broadcast

* Update redfire deps after RDS name change

* Remove RDS from redfire

Because several packages use redfire (which uses RDS) meaning the new
renamed RDS was a transitive dependency, but not yet in the github repo
Apparently that that combo is deadly for the pubspec_overrides mechanism
so I took RDS out of redfire. It's easy to put back in later