refactor(rdx): use endware for logic after actions
I added a DispatchEvent class, for now just for an endware in the
RedauxWidgetsTestUtils but it will probably also be used instead of
JsonMap for added type safety.

I changed parent to be a property of Action, as otherwise parent
would need to be a parameter of Endware which doesn’t seem right.

I made endwares for devtools and for test utils. The classes are
EmitDispatchEventsEndware and RecordDispatchEventsEndware.

I updated our_meals to the new redaux API.