Clean up and add a test
I added some more design docs, as well as moving some of the more 'hifi'
wireframes to Figma (as draw.io is great for lofi stuff but not so good for
the finer details). But I added a link to the Wireframe tab in the .dio
file so the Figma designs don't get lost.

I renamed action_history_item to actions_history_item so we are
consistent in the naming.

I started adding a test... and realised we need a redaux_test_utils
package like redfire_test but for redaux not redux :)