refactor(rdx): Rename Action members: leave/arrive
I renamed SyncAction.reduce to "arrive" and AsyncAction now has
"leave". The rename is meant to make it clearer that async actions go
off and do things elsewhere and generally end in a sync action causing a
state change.

I replaced AsyncAction.middleware with a function `leave` of type
Future<void> Function(Store<S>). The type parameter S comes from
the enlcosing class, ie. AsyncAction<S extends RootState>. The
reason is that this gives us the freedom to define the function in the
action declaration, or create a separate singleton then call the
singleton in the leave function... or have any other pattern we like.

In the Store I moved emitting state changes inside the chaeck for
the sync action type as there's no need to emit state changes on async
actions as they cannot change the state.