new(rxa): Add a redaux plugin package for auth (#171)
In this PR I updated packages using redaux to the new redaux API
from the last PR, pulled the auth related stuff out into a plugin
package and updated the packages to use the auth plugin.

Also I removed classes extending Middleware and put all the logic in the
`leave` function, with library-private globals for anything that needs
to live on after the function returns (eg. a Stream subscription). This
simplifies creating actions without reducing flexibility.

This is the start of implementing the idea of plugins (aka component
packages, still thinking on the name) by moving all of the auth related
stuff into a separatate redaux_auth package and importing the package
in both our_meal and resourganizer_desktop.

The plugin architecture is quite clunky at this stage but I will change
how it works to use source gen soon and I think that will be much better