chore(acf): Move authorization-code-flow to servers
The merged cloud-mono repo had a shelf server that was the start of a server for cloud run that could run an auth flow of the
authorization code type. There's not much in there yet but the next
step is to merge in old cloud-functions code from TheProcess that
managed this kind of flow.