chore(dip): move discord_interaction_to_pubsub_mes
I moved the discord_interaction_to_pubsub_message repo out of the
cloud-mono folder, which is from the recently merged cloud-mono repo.
The discord_interaction_to_pubsub_message repo is part of the
run-dart-code-on-discord solution which is not really being used but is
a good example of setting up quite a few different pieces into a useful
cloud based solution with Dart microservices communicating via pubSub
messages.