chore(+): Add AstroBase widget
I added an AstroBase widget with MissionControlProvider, MaterialApp,
AstroInspectorScreen and plugins, to our_meals & resourganizer_desktop.
The AstroBase widget will be generated in the future, probably with
TheProcess and/or an astro CLI tool.

The AstroBase widget holds most of the app-wide customisable parts
of the app which will be generated in the future, so this is a first step
in that direction.

The AstroBase widget includes a check to put the AstroInspectorScreen
in the widget tree iff the relevant dart-define has been set.

I added an "initAstroPlugins" function to astro_base.dart, which
currently just calls authInit(). When AstroBase is generated, all of
the plugins that have an init function will be called in the
initAstroPlugins. The user will have to call initAstroPlugins() in main
before runApp.

I couldn't resist and did a little bit of clean up and refactoring
(just naming) while I was wandering around the parts of the code
described above.