refactor(+): Add auto-parenting for away missions (#183)
I added AwayMissionControl which decorates MissionControl with a parent
mission. When MissionControl calls AwayMission.flightPlan it creates and
passes in an AwayMissionControl object with launch and land functions
that set the mission’s parent before calling MissionControl's
land/launch.

I renamed dockingMission to landingMission as we have landed (ay!) on
LandingMission/AwayMission for the naming.

I also replaced the type parameter S with T for conistency, as we were
using both in different places.