chore(+): Send mission upates to devtools plugin
I added a call to dart:developer's postEvent funciton in
astro_inspector_screen's SendMissionUpdatesToInspector system check,
sending the same mission updates as are already being sent to the in-app
inspector. That is all that is needed on the app side (at the other end
is our devtools plugin).

I fixed the AppState classes where I was mapping the errorMessages list
to json but wasn't returning the resulting iterable to a list.

I also did some more refactoring - renamed for clarity:
- OnMissionStartSystemCheck -> SendMissionUpdatesToInspector
- state-management folder -> missions folder
- missionEvents -> missionUpdates