new(aei): Add astro_error_handling_interface
This is a dart package that defines abstract classes to act as
interfaces so anyone can create their own astro error handling plugin.

We need a separate interface in this case as the error handling is
somewhat unavoidably baked into the code of the astro package.

So astro will depend on the interface, as will my plugin
'astro_error_handling' as will anyone that wants to make their own
plugin.