chore(aei): Update with BaseErrorHandlers mixin
I added a mixin that error handling plugins will use to
create a class that can be added to the MissionControl.

Currently the analyzer is happy but we get compiler errors
suggesting the interface types aren't available. I'm wondering
if it's due to a problem with the pubspec_overrides so this is also
a test to see if having the correct types in the main branch will fix
the problem.