Update interfaces
I added astro_core_interface so that plugins and the astro package
can both depend on the interfaces, and the plugins don't need to
depend on the astro package... almost like a kind of "inversion"
if you will.

I did some reading and decided to change the naming style to not use
Base for the interface and use Default for the implementation, the
interface types are what we should be using eveywhere so it makes more
sense to me this way.