General refactoring - ParentingMissionControl, etc
I renamed AwayMissionControl to ParentingMissionControl, made it
implemement MissionControl and changed the function signature of
AwayMission.flightPlan to take a MissionControl. This means we can
use a ParentingMissionControl as desired but can also use a regular
MissionControl any other time.

I renamed ErrorMessage to ErrorReport so we have report.message and
report.stack.

I moved files in astro package inside lib/src so we can selectively
export and added appropriate exports to lib/astro.dart.

I updated dependencies in our_meals to get new firebase_auth as latest
was broken.

I added an ErroPage to astro_error_handling, working toward a popup that
is shown when ErrorReports are added to AppState.