Update our_meals
I updated the code snippets and added snippets for Page and PageState.

I updated the our_meals package to use astro_types and the
updated astro_navigation.

I also added the astro_navigator package and a PagesNavigator widget to
handle turning state into a list of MaterialPages for the Navigator and
I updated AstroBase to use the new setup (ie. PagesNavigator).

I updated AppState to include the DefaultNavigationState members.