Move reports into an ErrorHandlingState member
I added a class ErrorHandlingState (with member List<ErrorReport>
reports), and and added member ErrorHandlingState error to
DefaultErrorHandlingState so we will access reports with
state.error.reports. The purpose is for readability and adds a class
for the errorhandling section so it will be easier to add members
later as desired.