feat(anv): Add missions & state for navigation (#198)
* Move interface packages into a single astro_types

* Remove redundant imports from redfire

* Update inspector & test utils to use astro_types

* Make astro_locator error message more helpful

* Add missions & required state to astro_navigation

I updated astro_navigation to use astro_types as the separate interface
packages have been removed.

I added a PushRoute landing mission that adds a given PageState to the
state's navigation stack, along with a RemoveCurrentRoute that pops the
top PageState off.

I added an extra layer of state so we have state.navigation.stack which
fits our section->feature structure and Navigation makes sense as a
section.

* Update astro_error_handling to astro_types

* Update astro_auth to astro_types & add route-sets

I added a route-set for an AuthCheck route, which will be put on the
navigation stack when the app loads and drive the authentication
sequence.

I also added a route-set for a SignIn route that will be pushed onto
the navigation stack when the auth check finds no credentials.

* Update astro to astro_types & add tests

I added some widgets for displaying progress with a message as it is a
common requirement and will probably be used in various user stories.

I started adding bunch of tests for MissionConttrol, AppState
OnStateChangeBuilder, etc. testing how state-based vs. identity
equivalence works at various points in the state management sequence.

* Update resourganizer_desktop

I updated the resourganizer_desktop package to use astro_types and the
updated astro_navigation.

I also added the astro_navigator package and a PagesNavigator widget to
handle turning state into a list of MaterialPages for the Navigator and
I updated AstroBase to use the new setup (ie. PagesNavigator).

I updated AppState to include the DefaultNavigationState members.

* Update our_meals

I updated the code snippets and added snippets for Page and PageState.

I updated the our_meals package to use astro_types and the
updated astro_navigation.

I also added the astro_navigator package and a PagesNavigator widget to
handle turning state into a list of MaterialPages for the Navigator and
I updated AstroBase to use the new setup (ie. PagesNavigator).

I updated AppState to include the DefaultNavigationState members.

* Fix astro_inspector_screen

I added a check in AppStateView for when a previous json list is
shorter than the current list, so we don't try and access above
where we can't.

* Add state-based equivalence to models

The models are AppState from our_meals, UserState from astro_auth &
NavigationState from astro_navigation. I'm doing this for now to remove
a potential cuase of bugs while continuing to look into different
strategies, eg. copy-on-write with identity equivalence.

* Change to single AuthGate over multiple routes

I changed to a single AuthGate widget that builds either a SignInScreen
or a HomeScreen, rather than using the navigator to push/pop routes.

* Remove problematic AuthCheck route-set

* Add ThrowingErrorHandlers to astro_error_handling

The handlers just throw again after catching - this is for use during
development where can be useful to have the debugger break on errors
rather than catch and handle them.

I also did a bit of organising, moving classes out of
astro_error_handling into separate files.

* Delete debug print statements

* Move reports into an ErrorHandlingState member

I added a class ErrorHandlingState (with member List<ErrorReport>
reports), and and added member ErrorHandlingState error to
DefaultErrorHandlingState so we will access reports with
state.error.reports. The purpose is for readability and adds a class
for the errorhandling section so it will be easier to add members
later as desired.

* Add a class for the auth section of the app state

I wrapped AppState.UserState in a new class: AuthState. This means
access becomes state.auth.user.   The purpose is for readability and
adding a class for the auth section so it will be easier to add members
later as desired.