feat(ast+): Push error reports on navigation stack (#201)
I registered the error-report route stack with the PageGenerator in the
locator so the PagesNavigator can generate an ErrorReportPage from
an ErrorReportPageState if such an object appears in the
state.navigation.stack.

I did some cleaning up and moved PagesNavigator to astro_navigation
which required making astro_navigation depend on astro.  Initially I
thought that was a bad idea but I made a dependency graph for the astro
packages we have so far and having plugins depend on astro doesn't
create any circular dependencies so that was a win.

I made AstroBase no longer take a child widget, as (for now) we will use
the astro_navigator package and put the initial screen up via a
route-set (Page, PageState, Screen).

I did some organising in astro_error_handling into renamed folders for
clarity and fixed the default error handlers & remove error mission to
deal with both error reports (in state.error.reports) and the
corresponding error report route (in state.navigation.stack). This
tightly couples astro_error_handling & astro_navigation but I plan to
rename them to something like default_error_handling &
default_navigation to make it clearer that the strategy is just one
option.