chore(*): Add caching of Flutter SDK to CICD (#205)
After a bit of fiddling I got this to work such that the SDK is being cached across workflow runs, per package as part of a matrix (runs a job per package in parallel). 

The cache key uses a hash of the current flutter stable version number. In a future PR I will add a check that the text file matches the actual stable version, print an error message and exit the workflow.

A subsequent CI run with cached Flutter SDKs was faster but not significantly so, next step is to get an average over time. My feeling is it's not going to make a huge difference but caching the package dependencies could be a bigger win.

* chore(*): Add caching of Flutter SDK to CICD

* Fix path, which doesn't seem to like $HOME

* Update the cache key to check cache busting

* Add package name to cache key

* Delete whitespace to kickoff again via PR

* Always put flutter on the path