new(iap): Add identify_affected_packages package (#207)
This is a Dart CLI app that will be used in CI to identify the set of
packages affected by the PR and should thus be tested. The app reads
in the set of all packages (saved in .github/matrix.json), calls git diff via
Process.run to get the paths of the changed files, checks those paths
against the set of all packages and writes out an adjusted matrix.json
so that the tests still run in parallel.