Complete migration - various fixes & refactoring
I fixed several model's toJson to have correct state_ members.

I added fromDocument functions to models that will be created from
Firestore documents. This couples the models and the
firestore_service_interface so I've made a task to consider decoupling,
maybe via extension methods.

I added calls to TapOrganisations in onInit and onDispose of the
OrganisationsSelector widget so the widget has the current data while
in the widget tree.

I added a FirestoreServiceFlutterfire object to the Locator under the
FirestoreService type so the service can be accessed with
locate<FirestoreService>() which is cleaner and easier to read.

I moved some code into new files under appropriate folders for clarity
and maintainability.

I removed code that was sending mission updates to any listening
DevTools as there is a bug and we're not currently using the feature. I've
added a task to fix the bug and put the feature back.

I fixed missions in astro_navigation that were not calculating new state
correctly.

I added a test that NavigationState json encodes correctly as a starting
point for testing all models to determine why sending to DevTools is
broken.