Add RecordingMissionControl etc. to astro_test_utils
I also added a Locator and added the default (overridable) option of
adding the MissionControl to the Loctor to WidgetTestHarness.

RecordingMissionControl just wraps a MissionControl, along with a
RecordMission system check, which is added to the MissionControl,
and exposes the funcitonality of the RecordMission system check.
Basiclly a convenience wrapper for tests allowing widgets etc. to check
whether given Missions were landed or launched.