chore(+): migrate the_process from redfire to astro (#225)
* Get old config running as a starting point

I ran flutterfire config to get missing files and followed the redfire
docs to create redfire_options.dart. The app runs on macos so we
have a working starting point.

* First attempt at migrating to astro

* Add missing toJson implementations

* Re-create macos dir to update generated code

* Update launch.json, commit updated .metadata

* Fix AstroBase to setup Inspector mode properly

We now only add the SendMissionUpdatesToInspector system check when the
'IN-APP-ASTRO-INSPECTOR' dart-define is found.

* Update snippets in our_meals

The snippets in our_meals are currently the default for astro as I
haven't figured out where to put them yet.

* Some small refactors and cleanups

* Delete cloud_firestore_extensions

The extensions have been moved to firestore_service_flutterfire as
currently the only time we use cloud_firestore is in
firestore_service_flutterfire so there is no need for a separate package.

* Delete redfire's firestore_service_flutterfire.dart

As the code has been moved to it's own separate package.

* Delete .flutter-plugins-dependencies

The file was commited but is a generated file not meant to be checked in.

* Gitignore generated file

Ignore .flutter-plugins-dependencies in firestore_service_flutterfire

* Fix launch.json to work in multiroot workspaces

* Complete migration - various fixes & refactoring

I fixed several model's toJson to have correct state_ members.

I added fromDocument functions to models that will be created from
Firestore documents. This couples the models and the
firestore_service_interface so I've made a task to consider decoupling,
maybe via extension methods.

I added calls to TapOrganisations in onInit and onDispose of the
OrganisationsSelector widget so the widget has the current data while
in the widget tree.

I added a FirestoreServiceFlutterfire object to the Locator under the
FirestoreService type so the service can be accessed with
locate<FirestoreService>() which is cleaner and easier to read.

I moved some code into new files under appropriate folders for clarity
and maintainability.

I removed code that was sending mission updates to any listening
DevTools as there is a bug and we're not currently using the feature. I've
added a task to fix the bug and put the feature back.

I fixed missions in astro_navigation that were not calculating new state
correctly.

I added a test that NavigationState json encodes correctly as a starting
point for testing all models to determine why sending to DevTools is
broken.

* Move parent out of Mission to new ParentedMission

* Add AvatarMenuButton to astro_auth

* Remove redfire_lints after updating to enspyr_lints

* Add RecordingMissionControl etc. to astro_test_utils

I also added a Locator and added the default (overridable) option of
adding the MissionControl to the Loctor to WidgetTestHarness.

RecordingMissionControl just wraps a MissionControl, along with a
RecordMission system check, which is added to the MissionControl,
and exposes the funcitonality of the RecordMission system check.
Basiclly a convenience wrapper for tests allowing widgets etc. to check
whether given Missions were landed or launched.

* Update the_process missions to new toJson style

We don't need anything but name_ and state_ now that we are
wrapping the MissionControl and creating parented missions in the
inspector.

* Update to the new enspyr_lints, add test utils dep

* Remove tests that were moved or merged together

* Add generated files from a 'flutter pub upgrade'

* Fix adding mission control to Locator in inspector

Forgot the dam type parameter... again :(

* Add an error handling strategy that just rethrows

Update MissionControl to make _errorHandlers nullable and rethrow on
catching an error if no error handler was set...
actual useful stack traces oh my goodness!

Update AstroBase in the_process to add the DefaultErrorHandlers, so
users will see helpful error messages. Also pass in the
ParentingMissionControl constructor tear off so the inspector has extra
useful info like each mission's parent.

* Fix mission's toJson - use correct type for state_

I made sure all 'state_' values were typed appropriately so the casting
is happy.

I also renamed for readability - event to missionUpdate, as the event
is a mission update.

* Fix AvatarMenuButton

I got avatar menu button working properly again, now from its new home
in astro_auth. There was no need for a "general" ui plugin and we should
try and resist the urge to create such catch-all containers.

* Update the_process macos setup

I previously configured apple sign in but the config files got lost
accidentally when I re-created the macos folder trying to fix a bug.

* Add plugin members to MissionControl in the_process

I added DefaultErrorHandlers & ParentingMissionControl -
the explanation was in a previous commit message, where
I thought this change had been added to the commit.

* Rename photUrl to photoURL in astro_auth UserState

This matches our style guide which comes from Effective Dart, ie.
acronyms should (with a few exceptions) be capitalised.

* Fix SendMissionUpdatesToInspector

The SendMissionUpdatesToInspector system check now adds the id and type
for each mission to the json added to the stream controller it was passed
in. This means the json that the Inspector recieves has all needed info.

* Update tests

* Add enspyr_lints to the_process

* Fix errors from new stricter lints

* Remove the check from CO run_tests job

* Remove redfire from matrix.json to stop testing it

redfire is being hacked up and replaced with astro so no point keeping
the tests working.

* Fix matrix.json - the_process has no sourcegen now

* Skip tests that fail due to beign partly finished

* Remove names in workflow files

Trying to fix stuck workflow run, we will just use job ids and avoid
the name key from now on (based on some reading of the docs)